Added options for ignoring pages based on nice URL, and options for current item & parent item classes.
When generating a nav menu during development, there might be pages that you do not want to show up in the prototype navigation.

By adding the URL of these to the list "ignore_pages" in the options, these pages will be skipped.

There is also the option of providing your own class names to the current item and items that have sub-lists in them via the options. These can be configured with the "is_current_class" and "is_parent_class", respectively, in the options.